partioned the UI a little more
