+++ src/aac/ui/model_views/diagram_view.py
@@ -1,0 +1,41 @@
+from tkinter import TOP, PanedWindow, Canvas
+from tkinter.ttk import Notebook, Frame
+
+
+def get_diagram_view(parent_window: PanedWindow):
+    tabs_notebook = Notebook(parent_window)
+
+    diagram_tab = _get_diagram_tab_frame(tabs_notebook)
+    properties_tab = _get_properties_tab_frame(tabs_notebook)
+    imports_tab = _get_imports_tab_frame(tabs_notebook)
+
+    tabs_notebook.add(diagram_tab, text="Diagram")
+    tabs_notebook.add(properties_tab, text="Properties")
+    tabs_notebook.add(imports_tab, text="Imports")
+    return tabs_notebook
+
+
+def _get_diagram_tab_frame(root_notebook: Notebook) -> Frame:
+    diagram_frame = Frame(root_notebook)
+    diagram_frame.pack(fill="both", expand=True)
+
+    model_canvas = Canvas(diagram_frame)
+    model_canvas.pack(side=TOP, fill="both", expand=True)
+
+    # Replace these with the model diagram content
+    model_canvas.create_line(0, 0, 2000, 1129)
+    model_canvas.create_line(0, 960, 1700, 0)
+
+    return diagram_frame
+
+
+def _get_properties_tab_frame(root_notebook: Notebook) -> Frame:
+    properties_frame = Frame(root_notebook)
+    properties_frame.pack(fill="both", expand=True)
+    return properties_frame
+
+
+def _get_imports_tab_frame(root_notebook: Notebook) -> Frame:
+    imports_frame = Frame(root_notebook)
+    imports_frame.pack()
+    return imports_frame
